Inline the `woltlab_environment` type in XSDs
Not all PIPs support the “all” environment, notably the templateListener PIP.
Thus the legal list of environments is now inlined to make it explicit and also
improve the readability of the XSDs.

Supporting the “all” environment for the templateListener PIP is not really
useful, as template names are rarely shared between frontend and ACP. It's
simple enough to register an explicit listener for each environment in the rare
case all of (templateName, eventName, templateCode) match up.

Fixes #5619