Fix the binding of existing removal buttons
The previous implementation of 5.5 lazily listened for clicks on the entire list which was refactored in 6.0 to bind on the delete button.

The existing contents of the ACL is statically generated in the template, but the delete button was not bound.

See https://www.woltlab.com/community/thread/301553-rechtevergabe-bei-seiten-l%C3%A4%C3%9Ft-sich-nicht-r%C3%BCckg%C3%A4ngig-machen/