Adopt the page overlay based on `[aria-hidden]`
The page overlay must be adopted if there is at least one dialog open and must be released again when the last dialog closed. The previous implementation did not work flawlessly if dialogs were stacked and toggled in arbitrary order.

The mutation observer solves the issue because we do not keep track of the number of open dialogs but only if the overlay is visible at all.

See https://www.woltlab.com/community/thread/302088-sporadisch-auftredende-fehlerhafter-verlinkungsmodus-im-editor/