Import `google/diff-match-patch`
This is a rather old library found at https://github.com/google/diff-match-patch/blob/master/javascript/diff_match_patch.js which is exported to npm through a 3rd party.

We cannot use a proper npm dependency because the export is a CommonJS module. The file is an exact copy of the file from the original repository with the added `define()` wrapper plus the `exports.` in the second to last line.