Reliably cancel an in-flight request
The `_previousXhr` property was not reliably set previously and the current in-flight was not visible unless a new request was dispatched.

The updated control flow in ca409b33e7550595d3ac2b50bb14b179237c13c9 effectively allowed for the `_previousXhr` property to be removed entirely. Working with `_xhr` directly allows to abort an in-flight request without dispatching a new request. This also fixes a third bug hidden in `_finalize()` that could have erased a later in-flight request in some cases.

See https://www.woltlab.com/community/thread/305578-halb-fehler-halb-wunsch/