Align the details of the benchmark’s stacktrace with the exceptions
This is now in line with the exceptions and makes it clear where a call happens.

This was previously quite confusing because the line is where the call in the file (!) took place that invoked said method. The line refers to the callee and not to a line within the class being listed.