Stop randomizing the bootstrapper order in dev mode
This was intended to reveal issues caused by missing dependencies where calls would only work because packages were installed (and thus bootstrapped) in a particular order.

However, this has the side effect that everything that depends on the scripts are randomized on every call. This becomes annoying with things like the ACP menu where menu items of apps start shifting around. The same applies to any action where things sometimes legitimately do not have a fixed order like rebuild workers with common dependencies.

We cannot solve these issues without giving up the randomization which is also not a priority feature and was merely intended to aid in finding errors.