Overhaul sorting of label groups in the admin panel
Previously, `showOrder` was used for sorting in the admin panel. However, `showOrder` was only intended for sorting in the frontend and could led to chaotic sorting in the admin panel when you have a larger number of label groups in your installation.

Instead, alphabetical sorting is now used consistently and the description is always displayed to make the label groups easier to distinguish.