Overhaul the Font Awesome picker
The previous implementation relied on the legacy implementation of dialogs which causes a stacking issue when opened from a new dialog.

Another flaw was the abysmal accessibility because it was impossible to reach the icons with a keyboard navigation.

In addition, the new implementation no longer relies on an initialization method which was used in a prior iteration.